+++ src/constants.ts
@@ -201,7 +201,7 @@ export const DEFAULT_OPTIONS: IKawkahOptions = {
   logFormat: '{{event}} {{message}} {{ministack|parens|muted}}',
 
   // Array of enabled middleware.
-  middleware: ['minmax', 'coerce', 'extend', 'required', 'validator', 'demand', 'deny', 'aliases'],
+  middleware: ['minmax', 'coerce', 'extend', 'required', 'validator', 'demand', 'deny', 'demandIf', 'denyIf', 'aliases'],
 
   // map of colors used when outputting to console.
   styles: {